move to MSVC

+++ geometry/hitablelist.h
@@ -20,7 +20,7 @@ public:
 bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) {
     hit_record temp_rec;
     bool hit_anything = false;
-    double closest_so_far = t_max;
+    float closest_so_far = t_max;
     for (int i = 0; i < list_size; i++) {
         if (list[i]->hit(r, t_min, closest_so_far, temp_rec)) {
             hit_anything = true;
